+++ hhnk_threedi_tools/core/climate_scenarios/klimaatsommen_prep.py
@@ -6,7 +6,7 @@ import hhnk_research_tools as hrt
 import pandas as pd
 
 from hhnk_threedi_tools import Folders
-from hhnk_threedi_tools.core.result_rasters.grid_to_raster import GridToRaster
+from hhnk_threedi_tools.core.result_rasters.grid_to_raster import GridToWaterLevel
 from hhnk_threedi_tools.core.result_rasters.netcdf_to_gridgpkg import NetcdfToGPKG
 
 
@@ -134,8 +134,11 @@ class KlimaatsommenPrep:
         output_file = scenario_raster
 
         # Init calculator
-        with GridToRaster(**calculator_kwargs) as basecalc:
-            basecalc.run(output_file=output_file, mode=mode, min_block_size=self.min_block_size, overwrite=overwrite)
+        if mode == "MODE_WLVL":
+            with GridToWaterLevel(**calculator_kwargs) as basecalc:
+                basecalc.run(
+                    output_file=output_file, mode=mode, min_block_size=self.min_block_size, overwrite=overwrite
+                )
 
     def calculate_depth(
         self,
