+++ hhnk_threedi_tools/core/result_rasters/grid_to_raster.py
@@ -1,15 +1,14 @@
-"""
-Most functions taken and edited from pip package: threedidepth.
-Modified to have more flexibility in input and calculation
-"""
-
 from pathlib import Path
+from typing import Union
 
+import geopandas as gpd
 import hhnk_research_tools as hrt
 import numpy as np
-from geopandas import GeoDataFrame
-from osgeo import gdal
+import pandas as pd
+import xarray as xr
+from scipy.interpolate import LinearNDInterpolator
 from scipy.spatial import Delaunay
+from shapely.geometry import LineString, Polygon
 from threedidepth import morton
 from threedigrid.admin.constants import NO_DATA_VALUE
 
@@ -27,215 +26,325 @@ htt.GridToRaster since v2024.2. Please rewrite your code."
 
 
 class GridToRaster:
-    """Calculate interpolated rasters from a grid. The grid_gdf is
-    created using the class ThreediGrid. Which converts the NetCDF
-    into a gpkg.
+    """TODO Deprecated, remove in later release."""
 
-    It is possible to calculate the wlvl and wdepth.
+    def __init__(self, **kwargs):
+        raise DeprecationWarning(
+            "The GridToRaster class has been named to \
+htt.GridToWaterLevel and htt.GridToWaterDepth since v2024.x. Please rewrite your code."
+        )
+
+
+def get_boundary_points(
+    grid_gdf: gpd.GeoDataFrame, boundary: LineString, wlvl_column: str, no_data_value: float
+) -> gpd.GeoDataFrame:
+    """Een extra rij met gridpunten op het middenpunt van de edges aan de rand van het 3Di grid
+
+    Parameters
+    ----------
+    grid_gdf : gpd.GeoDataFrame
+        3Di grid in GeoDataFrame
+    boundary : LineString
+        Rand van het 3Di data-domein
+    wlvl_column : str
+        kolom in GeoDataFrame met waterstanden
+    no_data_value : float
+        no_data_value in 3Di grid
+
+    Returns
+    -------
+    gpd.GeoDataFrame
+        GeoDataFrame met extra grid-punten aan rand van 3Di grid met data
+    """
+    # select only polygons with data
+    gdf = grid_gdf[grid_gdf[wlvl_column] != no_data_value]
+
+    # select only polygons at, but inside boundary
+    gdf = gdf[gdf.intersects(boundary)]
+
+    # get grid corner and edge-center points as fast as we can (we think :-))
+    bounds_gdf = gdf.bounds  # yields minx, miny, maxx and maxy
+    bounds_gdf.loc[:, "x"] = gdf.centroid.x  # centre x
+    bounds_gdf.loc[:, "y"] = gdf.centroid.y  # centre y
+
+    # get 4 corners and 4 edge-centers x and y, preserving index
+    x = pd.concat(
+        [
+            bounds_gdf.minx,
+            bounds_gdf.minx,
+            bounds_gdf.minx,
+            bounds_gdf.x,
+            bounds_gdf.maxx,
+            bounds_gdf.maxx,
+            bounds_gdf.maxx,
+            bounds_gdf.x,
+        ]
+    )
+
+    y = pd.concat(
+        [
+            bounds_gdf.miny,
+            bounds_gdf.y,
+            bounds_gdf.maxy,
+            bounds_gdf.maxy,
+            bounds_gdf.maxy,
+            bounds_gdf.y,
+            bounds_gdf.miny,
+            bounds_gdf.miny,
+        ]
+    )
+
+    # create a point GeoDataFrame
+    pnt_gdf = gpd.GeoDataFrame(geometry=gpd.points_from_xy(x, y), index=x.index, crs=grid_gdf.crs)
+
+    # as we preserved index, we can add original waterlevel
+    pnt_gdf.loc[:, wlvl_column] = gdf[wlvl_column]
+
+    # drop rows with equal geometry + waterlevel
+    pnt_gdf.drop_duplicates(inplace=True)
+
+    # select only points that are on the boundary
+    pnt_gdf = pnt_gdf[pnt_gdf.buffer(0.1).intersects(boundary)]
+
+    # take the average water-level of all points at the same location
+    gdf = gpd.GeoDataFrame(pnt_gdf.groupby("geometry").mean().reset_index(), crs=grid_gdf.crs)
+
+    return gdf
+
+
+def get_points_grid(
+    grid_gdf: gpd.GeoDataFrame, poly: Polygon, wlvl_column: str, no_data_value: float
+) -> gpd.GeoDataFrame:
+    """Selecteren van punten binnen het 3Di grid met data en converteren naar punten
+
+    Parameters
+    ----------
+    grid_gdf : gpd.GeoDataFrame
+        3Di grid in GeoDataFrame
+    poly : Polygon
+        Polygoon waarmee 3Di gridcellen worden gefilterd
+    wlvl_column : str
+        kolom in GeoDataFrame met waterstanden
+    no_data_value : float
+        no_data_value in 3Di grid
+
+    Returns
+    -------
+    gpd.GeoDataFrame
+        GeoDataFrame met grid-punten binnen en aan rand van 3Di grid met data
+    """
+    # select the the polygons adjacent to polygons with data
+    gdf = grid_gdf[grid_gdf[wlvl_column] == no_data_value]
+    gdf = gdf[gdf.intersects(poly.boundary)]
+
+    # concat with polygons with data
+    gdf = pd.concat([gdf, grid_gdf[grid_gdf[wlvl_column] != no_data_value]])
+
+    # take the centroid and change nodata to NaN
+    gdf.loc[:, "geometry"] = gdf.centroid
+    gdf.loc[gdf[wlvl_column] == no_data_value, wlvl_column] = pd.NA
+
+    return gdf
+
+
+def get_delaunay_grid(grid_gdf: gpd.GeoDataFrame, wlvl_column: str, no_data_value: float) -> gpd.GeoDataFrame:
+    """Uitlezen van alle punten uit het 3DI grid die handig zijn voor interpolatie
+
+    Parameters
+    ----------
+    grid_gdf : gpd.GeoDataFrame
+        3Di grid in GeoDataFrame
+    wlvl_column : str
+        kolom in GeoDataFrame met waterstanden
+    no_data_value : float
+        no_data_value in 3Di grid
+
+    Returns
+    -------
+    gpd.GeoDataFrame
+        GeoDataFrame met alle grid-punten waarmee delaunay triangulatie mee gemaakt wordt
+    """
+    # we maken een buffer rond de gridcellen met data; dit begrenst de interpolatie
+    poly = grid_gdf[grid_gdf[wlvl_column] != no_data_value].unary_union
+    boundary = poly.boundary
+
+    # we voegen alle relevante punten samen tot 1 grid
+    gdf = pd.concat(
+        [
+            get_points_grid(grid_gdf, poly, wlvl_column, no_data_value),
+            get_boundary_points(grid_gdf, boundary, wlvl_column, no_data_value),
+        ]
+    )
+
+    return gdf
+
+
+def get_interpolator(
+    grid_gdf: gpd.GeoDataFrame, wlvl_column: str, no_data_value: float, reorder=True
+) -> LinearNDInterpolator:
+    """LinearNDInterpolator op basis van Delauny triangulatie.
+
+    https://docs.scipy.org/doc/scipy/reference/generated/scipy.interpolate.LinearNDInterpolator.html
+
+    Parameters
+    ----------
+    grid_gdf : gpd.GeoDataFrame
+        3Di grid in GeoDataFrame
+    wlvl_column : str
+        kolom in GeoDataFrame met waterstanden
+    no_data_value : float
+        no_data_value in 3Di grid
+    reorder : bool, optional
+        optie om threedidepth.morton.reorder uit te voeren, staat default op True
+
+    Returns
+    -------
+    LinearNDInterpolator
+        LinearNDInterpolator
     """
 
-    PIXEL_MAP = "pixel_map"
-    LOOKUP_WLVL = "lookup_wlvl"
-    INTERPOLATOR = "interpolator"
-    DELAUNAY = "delaunay"
+    # aanmaken delaunay grid
+    gdf = get_delaunay_grid(grid_gdf, wlvl_column, no_data_value)
 
-    def __init__(
-        self,
-        dem_path,
-        grid_gdf,
-        wlvl_column,
-    ):
+    # converteren naar numpy-array met punt_grid
+    points_grid = gdf.get_coordinates().to_numpy()
+
+    # inlezen waterstanden als numpy-array
+    wlvl = gdf[wlvl_column].to_numpy()
+
+    # herordenen volgens threedidepth (geen idee wat het doet)
+    if reorder:
+        points_grid, wlvl = morton.reorder(points_grid, wlvl)
+
+    return LinearNDInterpolator(Delaunay(points_grid), wlvl)
+
+
+"""Calculate interpolated rasters from a grid. The grid_gdf is
+created using the class ThreediGrid. Which converts the NetCDF
+into a gpkg.
+
+It is possible to calculate the wlvl and wdepth.
+"""
+
+
+class GridToWaterLevel:
+    def __init__(self, dem_path: Path, grid_gdf: gpd.GeoDataFrame, wlvl_column: str):
+        """Bereken waterstanden geinterpoleerd naar de resolutie van de DEM, gebruikmakend van Delaunay-interpolatie
+
+        Parameters
+        ----------
+        dem_path : Path
+            Path naar de DEM
+        grid_gdf : gpd.GeoDataFrame
+            3Di grid in GeoDataFrame
+        wlvl_column : str
+            kolom in GeoDataFrame met waterstanden
+        """
         self.dem_raster = hrt.Raster(dem_path)
-
         self.grid_gdf = grid_gdf
-        self.wlvl_column = wlvl_column  # Column to use in calculation
+        self.wlvl_column = wlvl_column
+        self.wlvl_raster = None
+        self._interpolator = None
 
     @property
-    def lookup_wlvl(self):
-        """
-        Return the lookup table to find waterlevel by cell id.
+    def interpolator(self):
+        if self._interpolator is None:
+            self._interpolator = get_interpolator(
+                grid_gdf=self.grid_gdf, wlvl_column=self.wlvl_column, no_data_value=NO_DATA_VALUE
+            )
+        return self._interpolator
 
-        Both cells outside any defined grid cell and cells in a grid cell that
-        are currently not active ('no data') will return the NO_DATA_VALUE as
-        defined in threedigrid.
-        """
-        try:
-            return self.cache[self.LOOKUP_WLVL]
-        except KeyError:
-            lookup_wlvl = np.full((self.grid_gdf["id"]).max() + 1, NO_DATA_VALUE)
-
-            lookup_wlvl[self.grid_gdf["id"]] = self.wlvl_raw
-            self.cache[self.LOOKUP_WLVL] = lookup_wlvl
-        return lookup_wlvl
-
-    @property
-    def delaunay(self):
-        """
-        Return a (delaunay, s1) tuple.
-
-        `delaunay` is a scipy.spatial.Delaunay object, and `s1` is an array of
-        waterlevels for the corresponding delaunay vertices.
-        """
-        try:
-            return self.cache[self.DELAUNAY]
-        except KeyError:
-            points_grid = np.array(self.grid_gdf.centroid.apply(lambda x: [x.x, x.y]).to_list())
-
-            # reorder a la lizard
-            points_grid, wlvl = morton.reorder(points_grid, self.wlvl_raw)
-            delaunay = Delaunay(points_grid)
-            self.cache[self.DELAUNAY] = delaunay, wlvl
-            return delaunay, wlvl.astype(float)
-
-    def _get_points_mesh(self, window):
-        """Create mesh grid points with coordinates every 0.5m with the input window.
-        Point are created in the centre of the cell (0.5)
-        Args:
-            indices (tuple): ((i1, j1), (i2, j2)) subarray indices
-        """
-        # i1=leftx, i2=rightx (or y.. not sure)
-        (j1, i1), (j2, i2) = (window[0], window[1]), (window[0] + window[2], window[1] + window[3])
-        # Create meshgrid with window bounds
-        local_ji = np.mgrid[i1:i2, j1:j2].reshape(2, -1)[::-1].transpose()
-        xstart, xres, b, ystart, c, yres = self.output_raster.metadata.georef
-
-        # 0.5*res gives a point in the centre of the cell.
-        return local_ji * [xres, yres] + [xstart + 0.5 * xres, ystart + 0.5 * -yres]
-
-    def block_calculate_delauney_wlvl(self, window):
-        """Return waterlevel array. of the given window
-
-        This uses both the grid layout from the constant level method and the
-        triangulation from the linear method.
-
-        Interpolation is used to determine the waterlevel for a result cell if
-        all of the following requirements are met:
-        - The point is inside a grid cell
-        - The point is inside the triangulation
-        - The sum of weights of active (not 'no data' nodes) is more than half
-          of the total weight of all nodes. Only active nodes are included in
-          the interpolation.
-
-        In all other cases, the waterlevel from the constant level method is
-        used.
-        """
-        # start with the constant level result
-        nodeid_block = self.nodeid_raster._read_array(window=window)
-        # start with the constant level result
-        # node_id_grid is 1d array of the node ids in the mesh grid
-        nodeid_arr = nodeid_block.ravel()
-        # the waterlevel is known per nodeid. This lookuptable gets the waterlevel
-        # per point in the mesh grid
-        level = self.lookup_wlvl[nodeid_arr]
-
-        # determine result raster cell centers and in which triangle they are
-        points_mesh = self._get_points_mesh(window)
-        delaunay, wlvl = self.delaunay
-        simplices = delaunay.find_simplex(points_mesh)
-
-        # determine which points will use interpolation
-        in_gridcell = nodeid_arr != 0
-        in_triangle = simplices != -1
-        in_interpol = in_gridcell & in_triangle
-        points_int = points_mesh[in_interpol]
-
-        # get the nodes and the transform for the corresponding triangles
-        transform = delaunay.transform[simplices[in_interpol]]
-        simplices = delaunay.simplices[simplices[in_interpol]]
-
-        # calculate weight, see print(spatial.Delaunay.transform.__doc__) and
-        # Wikipedia about barycentric coordinates
-        weight = np.empty(simplices.shape)
-        weight[:, :2] = np.sum(transform[:, :2] * (points_int - transform[:, 2])[:, np.newaxis], 2)
-        weight[:, 2] = 1 - weight[:, 0] - weight[:, 1]
-
-        # set weight to zero when for inactive nodes
-        nodelevel = wlvl[simplices]
-        weight[nodelevel == NO_DATA_VALUE] = 0
-
-        # determine the sum of weights per result cell
-        weight_sum = weight.sum(axis=1)
-
-        # further subselect points suitable for interpolation
-        suitable = weight_sum > 0.5
-        weight = weight[suitable] / weight_sum[suitable][:, np.newaxis]
-        nodelevel = nodelevel[suitable]
-
-        # combine weight and nodelevel into result
-        in_interpol_and_suitable = in_interpol.copy()
-        in_interpol_and_suitable[in_interpol] &= suitable
-        level[in_interpol_and_suitable] = np.sum(weight.astype(float) * nodelevel.astype(float), axis=1)
-
-        # Return interpolated mesh grid
-        return level.reshape(nodeid_block.shape)
-
-    def block_calculate_delauney_wdepth(self, window):
-        """depth=wlvl-dem"""
-        # Read/create input
-        wlvl_block = self.block_calculate_delauney_wlvl(window)
-        dem_block = self.dem_raster._read_array(window=window)
-
-        # Mask output
-        mask = (dem_block == self.dem_raster.nodata) | (wlvl_block == self.output_raster.nodata)
-        # mask = np.any([masks[i] for i in masks],0)
-
-        # Calculate depth
-        block_out = wlvl_block - dem_block
-
-        block_out[block_out < -0.01] = self.output_raster.nodata
-        block_out[mask] = self.output_raster.nodata
-        return block_out
-
-    def create_nodeid_raster(self):
-        """Create raster of nodeids with same res as dem."""
-        if not self.nodeid_raster.exists():
-            hrt.gdf_to_raster(
-                gdf=self.grid_gdf,
-                value_field="id",
-                raster_out=self.nodeid_raster.path,
-                nodata=0,
-                metadata=self.dem_raster.metadata,
-                datatype=gdal.GDT_Int32,
-                read_array=False,
+    def run(self, output_file, chunksize: Union[int, None] = None, overwrite: bool = False):
+        # level block_calculator
+        def calc_level(_, dem_da, interpolator):
+            # get x and y coordinates from dem_da
+            x, y = np.meshgrid(
+                dem_da.x.data,
+                dem_da.y.data,
             )
 
-    def run(self, output_file, mode="MODE_WLVL", min_block_size=1024, overwrite=False):
-        # Init rasters
-        self.output_raster = hrt.Raster(output_file)
-        self.nodeid_raster = self.output_raster.parent.full_path("nodeid.tif")
+            # interpolate levels to x and y coordinates
+            level = interpolator(x, y)
+            level = np.expand_dims(level, axis=0)
 
-        create = hrt.check_create_new_file(output_file=self.output_raster.path, overwrite=overwrite)
+            # Return as xarray DataArray, using dem_da's coordinates
+            level_da = xr.full_like(dem_da, dem_da.rio.nodata)
+            level_da.values = level
+
+            return level_da
+
+        # get dem as xarray
+        dem = self.dem_raster.open_rxr(chunksize)
+
+        # init result raster
+        self.wlvl_raster = hrt.Raster(output_file, chunksize=chunksize)
+        create = hrt.check_create_new_file(output_file=self.wlvl_raster.path, overwrite=overwrite)
 
         if create:
-            # Create rasters
-            self.output_raster.create(metadata=self.dem_raster.metadata, nodata=NO_DATA_VALUE)
-            self.create_nodeid_raster()  # Nodeid raster for calculation
+            # create empty result array
+            result = xr.full_like(dem, dem.rio.nodata)
 
-            self.wlvl_raw = np.array(self.grid_gdf[self.wlvl_column]).copy()  # wlvl list
+            result = xr.map_blocks(calc_level, obj=result, args=[dem, self.interpolator], template=result)
 
-            # Open outputraster
-            target_ds = self.output_raster.open_gdal_source_write()
-            target_band = target_ds.GetRasterBand(1)
+            self.wlvl_raster.write(output_file, result=result, nodata=dem.rio.nodata, chunksize=chunksize)
 
-            # Do calculations per block
-            for idx, window, block_row in self.output_raster.iter_window(min_block_size=min_block_size):
-                if mode == "MODE_WLVL":
-                    block_out = self.block_calculate_delauney_wlvl(window)
-                if mode == "MODE_WDEPTH":
-                    block_out = self.block_calculate_delauney_wdepth(window)
-
-                target_band.WriteArray(block_out, xoff=window[0], yoff=window[1])
-            target_band.FlushCache()  # close file after writing
-            target_band = None
-            target_ds = None
+        return self.wlvl_raster
 
     def __enter__(self):
         """With GridToRaster(**args) as x. will call this func."""
-        self.cache = {}
+        self._interpolator = None
         return self
 
     def __exit__(self, *args):
-        self.cache = {}
+        self._interpolator = None
+
+
+class GridToWaterDepth:
+    def __init__(self, dem_path: Path, wlvl_path: Path):
+        """Bereken waterstanden geinterpoleerd naar de resolutie van de DEM, gebruikmakend van Delaunay-interpolatie
+
+        Parameters
+        ----------
+        dem_path : Path
+            Path naar de DEM
+        wlvl_path : Path
+            Pad naar grid met waterstanden
+        """
+        self.dem_raster = hrt.Raster(dem_path)
+        self.wlvl_raster = hrt.Raster(wlvl_path)
+        self.depth_raster = None
+
+        if not self.wlvl_raster.path.exists():
+            raise FileNotFoundError(
+                f"{self.wlvl_raster.path} does not exist. Provide valid path or run GridToWaterLevel first."
+            )
+
+    def run(self, output_file, chunksize: Union[int, None] = None, overwrite: bool = False):
+        # level block_calculator
+        def calc_depth(_, dem_da, level_da):
+            depth_da = level_da - dem_da
+
+            depth_da = xr.where(depth_da < 0, 0, depth_da)
+
+            return depth_da
+
+        # get dem as xarray
+        dem = self.dem_raster.open_rxr(chunksize)
+        level = self.wlvl_raster.open_rxr(chunksize)
+
+        # init result raster
+        self.depth_raster = hrt.Raster(output_file, chunksize=chunksize)
+        create = hrt.check_create_new_file(output_file=self.depth_raster.path, overwrite=overwrite)
+
+        if create:
+            # create empty result array
+            result = xr.full_like(dem, 0)
+
+            result = xr.map_blocks(calc_depth, obj=result, args=[dem, level], template=result)
+
+            self.depth_raster.write(output_file, result=result, nodata=0, chunksize=chunksize)
+
+        return self.depth_raster
 
 
 if __name__ == "__main__":
@@ -257,10 +366,6 @@ if __name__ == "__main__":
     }
 
     # Init calculator
-    with GridToRaster(**calculator_kwargs) as self:
-        # self.run(output_file=threedi_result.full_path("wlvl_orig.tif")
-        #             mode="MODE_WLVL",
-        #             overwrite=OVERWRITE)
-
+    with GridToWaterLevel(**calculator_kwargs) as self:
         self.run(output_file=threedi_result.full_path("wdepth_orig.tif"), mode="MODE_WDEPTH", overwrite=OVERWRITE)
         print("Done.")
